Split `SocketEngine` into `tcp` and `udp` parts and remove contexts, as those are really just function calls.
