+++ ktor-network/web/src/io/ktor/network/sockets/SocketEngine.udp.web.kt
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2014-2026 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package io.ktor.network.sockets
+
+import io.ktor.network.selector.*
+
+internal actual suspend fun udpConnect(
+    selector: SelectorManager,
+    remoteAddress: SocketAddress,
+    localAddress: SocketAddress?,
+    options: SocketOptions.UDPSocketOptions
+): ConnectedDatagramSocket = error("UDP sockets are unsupported on WASM/JS")
+
+internal actual suspend fun udpBind(
+    selector: SelectorManager,
+    localAddress: SocketAddress?,
+    options: SocketOptions.UDPSocketOptions
+): BoundDatagramSocket = error("UDP sockets are unsupported on WASM/JS")
